More release work, builds can now launch 1.16 with proxy.
