+++ src/gg/codie/mineonline/URLSetAdvice.java
@@ -4,9 +4,9 @@ import net.bytebuddy.asm.Advice;
 
 import java.lang.reflect.Field;
 
-class URLAdvice {
+class URLSetAdvice {
     @Advice.OnMethodEnter
-    static void intercept(@Advice.Origin Object url, @Advice.Argument(1) String host) {
+    static void intercept(@Advice.Origin Object url, @Advice.Argument(0) String protocol, @Advice.Argument(1) String host) {
         try {
             if(!host.isEmpty())
                 System.out.println("HOST: " + host);
@@ -18,6 +18,7 @@ class URLAdvice {
 //            Settings.loadSettings();
 //            System.out.println("Settings: " + settings);
             for(String replaceHost : new String[] {
+                    "pc.realms.minecraft.net",
                     "www.minecraft.net:-1",
                     "skins.minecraft.net",
                     "session.minecraft.net",
@@ -33,15 +34,17 @@ class URLAdvice {
                     "sessionserver.mojang.com",
 
                     "banshee.alex231.com",
-                    "mcauth-alex231.rhcloud.com"
+                    "mcauth-alex231.rhcloud.com",
+                    "pc.realms.minecraft.net"
             }) {
 //                System.out.println(replaceHost);
-                if(host.equals(replaceHost)) {
+                if(host.contains(replaceHost)) {
                     Field f = String.class.getDeclaredField("value");
                     f.setAccessible(true);
                     f.set(host, host.replace(replaceHost, Globals.API_HOSTNAME).toCharArray());
+                    f.set(protocol, protocol.replace("https", "http").toCharArray());
 
-                    System.out.println("Replaced." );
+                    System.out.println("Replaced.");
                 }
             }
